Wrapper: - Refactor the setup of class loaders so that it is clearer what is   going on - For the Log4J tests, add the custom Appender implementation to   whichever class loader Log4J is put in - Set and reset the thread context class loader to more accurately   simulate the environment of a servlet container.
build.xml:
- Implement the multi-classloader test scenarios for Log4J, equivalent
  to those for JDK 1.4.

We now have a robust testing environment to catch any possible regressions
when modifying the functionality (such as ripping out Log4jFactory, as
discussed on COMMONS-DEV).

I don't have time to do the corresponding test scenarios for Avalon LogKit,
but anyone who wants to do so is welcome to create them.


git-svn-id: https://svn.eu.apache.org/repos/asf/jakarta/commons/proper/logging/trunk@138965 13f79535-47bb-0310-9956-ffa450edef68