Use maven-failsafe-plugin to run the tests, as this plugin is bound to the "integration-test" phase which comes after the "package" phase. This is what commons-logging needs for most of its tests.
The normal "test" phase is handled by maven-surefire-plugin. We "disable" that by only running a single test class, one that can be run at the "test" phase. This a hack. To do it the Maven way would mean renaming all integration test source files. But, because we still have an Ant build to maintain I prefer to make as few changes as possible.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/logging/trunk@1370536 13f79535-47bb-0310-9956-ffa450edef68